webapi/tct-notification-w3c-tests/notification/notification_attribute.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xinx, liu <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>Notification Test: notification_attribute</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-notifications-20120614/#api" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the notification attributes are implemented as specified for existence, type" />
    <script src="support/notifications.js"></script>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script type="text/javascript">
        getnotification();
        test(function() {
                assert_equals(typeof notification, "object", name + " notification supports ");
        }, "Check if notification supports");

        [
            ["eventhandler", "onclick"],
            ["eventhandler", "onshow"],
            ["eventhandler", "onerror"],
            ["eventhandler", "onclose"],
            ["Function", "close"]
        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];

            // existence check
            test(function() {
                assert_true(name in notification, name + " attribute exists ");
            }, "Check if notification." + name + " attribute exists");

            // type check
            switch(type) {
                case "eventhandler":
                    test(function() {
                        assert_equals(typeof notification[name], "object", name + " attribute of type");
                    }, "Check if the type notification." + name + " attribute is object");
                    break;
                case "Function":
                    test(function() {
                        assert_equals(typeof notification[name], "function", name + " attribute of type");
                    }, "Check if the type notification." + name + " attribute is function");
                    break;
                default:
                    break;
            }
        });
    </script>
</body>
</html>
